requirements_data_collector.txt
# Sofia V2 - Free Data Collection System Requirements
# Comprehensive package list for collecting crypto, BIST, and social data

# Core Web Scraping & HTTP
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.0.0
fake-useragent==1.4.0

# Cryptocurrency APIs and WebSockets  
ccxt==4.1.56
websockets==12.0
python-binance==1.0.19

# RSS and News Processing
feedparser==6.0.10
newspaper3k==0.2.8
python-dateutil==2.8.2

# Social Media (Free alternatives)
snscrape==0.5.0.20230113
telethon==1.31.1
tweety-ns==1.0.9.5

# Data Processing & Analysis
pandas==2.1.3
numpy==1.24.3
pytz==2023.3

# Database & Cache
redis==5.0.1
sqlalchemy==2.0.23
influxdb-client==1.38.0

# Task Scheduling & Background Jobs
apscheduler==3.10.4
celery==5.3.4

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2

# Optional: Advanced Features
# scrapy==2.11.0              # For large-scale scraping
# playwright==1.40.0          # For JavaScript-heavy sites  
# selenium==4.15.0            # Backup browser automation
# transformers==4.35.2        # For sentiment analysis
# torch==2.1.0               # PyTorch for ML models